pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qatis"
version = "0.1.0"
description = "QATIS: OSINT collection and AI evaluation pipeline (CLI + Streamlit UI)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  { name = "QATIS" }
]
dependencies = [
  "typer>=0.12.5",
  "python-dotenv>=1.0.1",
  "requests>=2.31",
  "PyYAML>=6.0",
  "openai>=1.30.0",
  "trafilatura>=1.7",
  "tqdm>=4.66",
  "tenacity>=8.4",
  "python-slugify>=8.0",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "snscrape==0.7.0.20230622",
  "yt-dlp>=2024.7.1",
  "telethon>=1.36",
  "scraperapi-sdk>=0.2.3",
  "pyalex>=0.13"
]

[project.optional-dependencies]
ui = [
  "streamlit>=1.37",
  "pandas>=2.0"
]

[project.scripts]
qatis = "qatis.cli:app"

[tool.setuptools]
packages = { find = { where = ["."] } }

[tool.setuptools.package-data]
qatis = ["prompts/*.md"]


